Fix delay when loading new service item
Often kernels can't stop a program for such a specific/small amount of time.
On my system I was encountering half second delays with this sleep statement.
Most uses of wait_for require some other part of Qt to do stuff so sleeping
is not a good choice in many cases.